Use --current-version with --generate-api-version-history
Previously, when `--generated-api-version-history` was specified the
current version had to be provided as the last item in the
`--api-version-names` option. This change allows it to be provided as
`--current-version` instead. This is preparation for avoiding having to
provide the `--api-version-names` and `--api-version-signature-files`.
Instead it will use the pattern matching to extract the versions from
the file names.

Bug: 388441272
Test: ./gradlew
Change-Id: I64a432f20e01853e23623028b7156b7751dcb4b9