Add FileProvider abstraction to access files to scan
Previously, `PatternNode.scan(...)` had access to all the `File`s in
the file system. However, AndroidX passes in a list of files that it
wants to contribute to the API history as it will only use one file for
each `major.minor.patch`. e.g. If it has two files with different
quality like `1.1.1-beta01` and `1.1.1-alpha01` it will use the version
with the highest quality, i.e. `1.1.1-beta01`. Similarly, if it has one
file with a quality and one without it will use the one without as that
is the released version.

There are two ways to handle this:
1. Iterate over the list of files matching each against the
   `PatternNode` hierarchy. The problem with this approach is that the
   patterns do not drive the behavior like they do when scanning and it
   requires each `PatternNode` implementation to add support for
   matching.

2. Using the list of files to create a limited view of the file system
   that only contains the listed files (and their containing
   directories) and then scanning that limited view.

The second approach of limiting access to the file system but reusing the
scanning code will be used as it is simpler and maintains the behavior of
the patterns, i.e. the patterns drive the scanning and determined the
order of the results.

This change adds a `FileProvider` abstract for use by `PatternNode`s
that need to access the file system. It defaults to an implementation
that provides access to the whole file system. A follow-up change will
add a limited view of the file system.

Bug: 388441272
Test: ./gradlew
Change-Id: I80b97c1c998c2ea3b2773c1d2c3cac654e53018f