Correct unrealistic ExtractSystemApiLevelsTest
Previously, it would actually read all the extension jars for all
API surfaces, e.g. `public`, `system`, `module-lib`, etc. That is not a
problem in real use because the Android build system uses a sandbox to
limit Metalava's access to only those extension jars in the correct
surface, e.g. in `prebuilts/sdk/extensions/<version>/system` when
generating the system API.

This change adds a new `--api-surface` command line option to allow the
API surface currently being generated to be supplied. That is then
passed through to `ExtensionSdkJarReader` to use to limit access to
only those jars for the correct surface. If it is not provided then it
falls back to the previous behavior of reading any available jar. That
is needed because the Android build system does not yet specify the
`--api-surface` command line option.

That revealed an issue with the test as it could not find the
`setBrowserRoleHolder(...)` line. That is because it is part of the
`module-lib` API surface. See https://cs.android.com/android/platform/superproject/main/+/main:packages/modules/Permission/framework-s/java/android/app/role/RoleManager.java;l=948;drc=4675be8559d10d1fb60bfde9bfe4214ec5ad2c64.

The test was modified to use a different method with similar
characteristics, i.e. an `sdks` attribute different to its parent's.

A future change will make the Android build system supply the
`--api-surface` command line option allowing the legacy behavior to be
removed.

Bug: 259115852
Test: ./gradlew
Change-Id: I406ed823ebd9213d328aa189930a8f2209f209e6