Enhance BaseOptionGroupTest to test ApiLevelsGenerationOptions
Previously, it was impossible to test the `forAndroidConfig(...)` in
`ApiLevelsGenerationOptionsTest` as it accesses the
`EarlyOptions.verbosity` option property which is not initialized by
the test. This change adds support to allow an `OptionGroup` class to
declare `OptionGroup`s it requires which are then initialized by
`BaseOptionGroupTest`, if required.

While the required `OptionGroup`s are included by default the
`Test help` test method purposely does not include them as otherwise
they would be included in the generated help.

A follow-up change will use this mechanism to add tests to the
`ApiLevelsGenerationOptionsTest` class for `forAndroidConfig(...)`.

Bug: 259115852
Test: ./gradlew
Change-Id: I2ebc2a21f66a2bf0362666182eaea208f0b9f70a