Verify forAndroidConfig(...) handles extensions correctly
Adds a test to verify that the `forAndroidConfig(...)` handles
finding extension jars correctly. This is in preparation for follow-up
changes that will refactor the code and merge the finding of extension
jars into the code that scans for `android.jar` files which will end
up removing the `ExtensionSdkJarReader.findExtensionSdkJarFiles()`
method and require removing the test.

Adding this test now ensures coverage of the code is not affected by
the refactoring.

Bug: 259115852
Test: ./gradlew
Change-Id: I4f5a3363f56887c42e847b68122a1a22d9049211