Use version 4 of api-versions.xml for major.minor versions
Previously, Metalava would generate `api-versions.xml` file with
`<api version="3"..>`. If the file contains `major.minor` versions then
it needs to use version 4 so that the consumers, e.g. Android Lint
`ApiLookup` class, knows to expect `major.minor` versions.

However, not all consumers will be using a version of `ApiLookup` that
supports so it should continue to use version 3 for as long as possible
to give them time to upgrade and avoid any disruption.

So, this change makes Metalava use version 4 only if the generated file
contains `major.minor` versions. As none of the existing tests use
`major.minor` versions (some do use `major.minor.patch`) this change
adds a test that uses `major.minor` versions to show that it does use
version 4 in that case.

Bug: 302189716
Test: ./gradlew
Change-Id: I02ebbf39c723df0f4b0ed5fbe5d0983eebcd2aee