Map annotations from typealiases to the aliased type
Using K2, an expect annotation in common with an actual typealias in android/jvm is represented with the expect annotation type, not the value of the actual typealias like it is with K1. In the case of @JvmDefaultWithCompatibility, the value of the typealias is important for API compatibility. This collects the typealiases present in a codebase and while creating annotations checks if the annotation name is a typealias.

Bug: 384619031
Test: JvmDefaultWithCompatibility case in UastTestBase
Change-Id: Ic7c392a75a0e90a2c0b65ba5f6abb4ea1bc10b72