Create ApiSurface objects from ApiSurfaceConfig
Previously, when a target `ApiSurfaceConfig` was selected the
`ApiSurfaces` (and corresponding `ApiSurface`) objects were created
solely based on whether its `ApiSurfaceConfig.extends` property was
`null` or not. This change used the whole extends hierarchy of the
target `ApiSurfaceConfig` and their names when creating them.

The `ApiSurface` instances are created in order from extended surfaces
to extending surfaces. e.g. `public` comes before `system` because
`public` is extended by `system`.

Bug: 391554590
Test: ./gradlew
Change-Id: I63edfad0af9625d4bfff63d905db2a022831716d