Add ApiSurfaces.Builder and ApiSurfaces.build()
Previously, an `ApiSurfaces` object was created using the
`ApiSurfaces.create(needsBase:Boolean)` method which only allowed
control over whether there was one `ApiSurface` or two, one which
extends the other. This change reimplements that in terms of a new
`build(...)` method and `Builder` class.

It works similar to the previous code in that the creation of the
`ApiSurface` instances is done in the initializer of the
`DefaultApiSurfaces` instance. The main difference is that instead of
that behavior being hard-coded into the `DefaultApiSurfaces` class it
is now provided as a lambda that takes a `Builder` context object.

The `Builder` interface defines a single method `createSurface(...)`.
The `DefaultApiSurfaces` provides a `BuilderImpl` implementation of
that which creates and keeps track of the `DefaultApiSurface` instances
and is then used to initialize the `DefaultApiSurfaces` properties.

The previous hard-coded behavior is reimplemented in
`simpleSurfacesInitializer(needsBase)` and used by the
`create(needsBase)` method.

A follow-up change will use that to create `ApiSurfaces` from
`ApiSurfacesConfig`.

Bug: 391554590
Test: ./gradlew
Change-Id: I7fccec83b74602fea032d4b8da5ce849e0e5c622