Add ApiSurfacesConfig.getByNameOrError(...)
Preparation for follow-up changes that look up `ApiSurfaceConfig` by
name and fail if it could not be found. Allows each caller to customize
the error message.

This changes the type of exception from `MetalavaCliException` to
`IllegalStateException` and while that requires a minor unit test
change it does not impact other tests or in practice as the latter
exception is converted into the former further up the stack.

Bug: 391554590
Test: ./gradlew
Change-Id: I5d3fce67c532edb1c5fadb0fe78e44ef66be104b