Make PatternNodeTests easier to write
Previously, the tests compared lists of `MatchedPatternFile`. That
necessitated creating all the correct objects and when tests failed
all that was returned was a difference between two long lines of the
contents of that list. That made it difficult to see the differences
and provided no support for creating the correct input, e.g. for new
tests.

This change switches the tests to compare the string representation of
the lists, formatted one item per line with the string representation
optimized to ignore optional properties. When tests fail the actual
output can just be copied directly into the test (assuming it is
correct).

Bug: 378479241
Test: ./gradlew
Change-Id: I0a5219fb6b77fca572522025b0166fb6c6a9400a