Pass VersionedApis to ApiXmlPrinter instead of ApiVersions
Previously, an `ApiXmlPrinter` was passed in an `ApiVersion` list that
it used to compute the version in which an API was removed based on
when it was last present. Sometimes, there was an inconsistency between
the versions in that list and the versions of the `VersionedApi`s that
were used to create the `Api` that the printer was printing. That lead
to exceptions.

This reduces the potential for inconsistencies by passing in the
`VersionedApi`s and computing the `ApiVersion` list from that directly.

Bug: 383521289
Test: ./gradlew
Change-Id: Ie9db0b26374bf789ff7a5635d7fe44ee1b6f3a3b