Ensure consistent naming in Api
Previously, the decision as to whether to use internal names or source
names in the `Api` object was determined by the `VersionedApi`
implementations:
* `VersionedJarApi` always used internal names.
* `VersionedSourceApi` was configurable.
* `VersionedSignatureApi` never used internal names.

That meant it was possible to use an inconsistent mixture of
`VersionedApi` implementations. This change moves the decision as to
whether internal name are moved into the `Api` object itself. As
`VersionedJarApi` can only use internal names it checks that `Api`
uses them too before attempting to update it.

There was no point adding support for source names to `VersionedJarApi`
as that will be removed once `--generate-api-levels` supports signature
files and there are no plans to use it with
`--generate-api-version-history`.

Bug: 378479241
Test: ./gradlew
Change-Id: Ie1144ec24777a49f683439f4673c0895f61c4282