Sort ApiJsonPrinter output
Previously, `ApiJsonPrinter` did not sort the methods and fields before
printing which meant the contents of the file was sensitive to
implementation details (e.g. like the type of map used to store them
in `ApiClass`) and while it appears to be stable there are no
guarantees.

`ApiClass` used `ConcurrentHashMap` to store the methods and fields.
That was added in https://r.android.com/2134184 and it is not explained
why that was used but it does not appear to have been for thread safety
but rather because the iterators do not throw
`ConcurrentModificationException`. The order of iteration of that is
undefined and different to `mutableMap()` which uses a `LinkedHashMap`
and so order is defined.

This change switches the map from `ConcurrentHashMap` to `mutableMap()`
which broke some `ApiJsonPrinter` based tests. They were fixed by first
sorting methods and fields in `ApiJsonPrinter` and then updating the
tests to match.

It was considered to use `TreeMap` for methods and fields but as the
items only need to be sorted once it was not worth the increased costs
of insertion (O(logN) for `TreeMap` vs O(1) for `LinkedHashMap`).

Bug: 379646963
Test: ./gradlew
Change-Id: I9bba191c9050b6db61564e6ebf1e0e095a4c7281